Fix Bug in 'Word Chain' CodeKata
The problem was, that I thought, the character difference to the target word
should always be the lowest. This assumption is false, because it can lead to
a dead end road. To fix the problem, now I use the lowest difference to the
target word AND the level above (e.g. all words with difference of 3 characters
AND 4 characters to the target word (for tree level X)). I also try to memorize
the words I have already seen, to minimize the search space. It works, but I'm
not completely happy with it: (rust -> java takes 5s) :/